requirements.txt

python-dotenv
langchain_openai
langchain_ollama
transformers 
langchain-anthropic